Rename 'EuroPythonBot' to 'europython_discord'
