+++ .github/workflows/deploy.yml
@@ -26,7 +26,7 @@ jobs:
         with:
           ssh-private-key: ${{ secrets.DISCORD_BOT_DEPLOY_KEY }}
 
-      - name: Deploy EuroPythonBot with ansible
+      - name: Deploy Discord Bot with ansible
         run: ansible-playbook -i ansible/inventory.ini ansible/deploy-playbook.yml --private-key="/home/runner/.ssh/id_rsa" --user=root
         env:
           ANSIBLE_HOST_KEY_CHECKING: "false"
