+++ .gitignore
@@ -3,7 +3,7 @@
 .idea
 .vscode
 README.md
-EuroPythonBot/config.local.toml
+src/europython_discord/config.local.toml
 __pycache__
 .DS_Store
 registered_log.txt
