+++ src/europython_discord/bot.py
@@ -8,10 +8,10 @@ import discord
 from discord.ext import commands
 from dotenv import load_dotenv
 
-from EuroPythonBot import configuration
-from EuroPythonBot.cogs.ping import Ping
-from EuroPythonBot.program_notifications.cog import ProgramNotificationsCog
-from EuroPythonBot.registration.cog import RegistrationCog
+from europython_discord import configuration
+from europython_discord.cogs.ping import Ping
+from europython_discord.program_notifications.cog import ProgramNotificationsCog
+from europython_discord.registration.cog import RegistrationCog
 
 load_dotenv(Path(__file__).resolve().parent.parent.parent / ".secrets")
 DISCORD_BOT_TOKEN = os.getenv("DISCORD_BOT_TOKEN")
@@ -76,7 +76,7 @@ async def run_bot(bot: Bot) -> None:
         await bot.add_cog(Ping(bot))
         await bot.add_cog(RegistrationCog(bot))
         await bot.add_cog(ProgramNotificationsCog(bot))
-        await bot.load_extension("EuroPythonBot.extensions.organisers")
+        await bot.load_extension("europython_discord.extensions.organisers")
         await bot.start(DISCORD_BOT_TOKEN)
