+++ src/europython_discord/configuration.py
