+++ src/europython_discord/program_notifications/models.py
