+++ src/europython_discord/program_notifications/program_connector.py
@@ -7,7 +7,7 @@ from pathlib import Path
 import aiofiles
 import aiohttp
 
-from EuroPythonBot.program_notifications.models import Break, Schedule, Session
+from europython_discord.program_notifications.models import Break, Schedule, Session
 
 _logger = logging.getLogger(f"bot.{__name__}")
