+++ src/europython_discord/program_notifications/session_to_embed.py
@@ -5,7 +5,7 @@ from typing import Final
 
 from discord import Embed
 
-from EuroPythonBot.program_notifications.models import Session, Speaker
+from europython_discord.program_notifications.models import Session, Speaker
 
 _AUTHOR_WIDTH: Final = 128
 _TWEET_WIDTH: Final = 200
