+++ src/europython_discord/registration/ticket.py
