+++ tests/program_notifications/test_program_connector.py
@@ -9,7 +9,7 @@ import pytest
 from aiohttp import web
 from aiohttp.test_utils import TestServer
 
-from EuroPythonBot.program_notifications.program_connector import ProgramConnector
+from europython_discord.program_notifications.program_connector import ProgramConnector
 
 mock_schedule_file = Path(__file__).parent / "mock_schedule.json"
